+++ src/citra_qt/multiplayer/host_room.cpp
@@ -1,4 +1,4 @@
-// Copyright 2017 Citra Emulator Project
+// Copyright Citra Emulator Project / Azahar Emulator Project
 // Licensed under GPLv2 or any later version
 // Refer to the license.txt file included.
 
@@ -142,8 +142,8 @@ void HostRoomWindow::Host() {
         }
         if (auto room = Network::GetRoom().lock()) {
             bool created = room->Create(ui->room_name->text().toStdString(),
-                                        ui->room_description->toPlainText().toStdString(), "", port,
-                                        password, ui->max_player->value(),
+                                        ui->room_description->toPlainText().toStdString(), "",
+                                        static_cast<u16>(port), password, ui->max_player->value(),
                                         NetSettings::values.citra_username, game_name.toStdString(),
                                         game_id, CreateVerifyBackend(is_public), ban_list);
             if (!created) {
@@ -193,7 +193,8 @@ void HostRoomWindow::Host() {
         }
 #endif
         member->Join(ui->username->text().toStdString(), Service::CFG::GetConsoleIdHash(system),
-                     "127.0.0.1", port, 0, Network::NoPreferredMac, password, token);
+                     "127.0.0.1", static_cast<u16>(port), 0, Network::NoPreferredMac, password,
+                     token);
 
         // Store settings
         UISettings::values.room_nickname = ui->username->text();
